test: add fixture base, split test suites, and split Kotlin route collector tests
- Introduce ArmeriaFixtureTestBase and ArmeriaClientFixtureTestBase to deduplicate
  Armeria PSI stubs across fixture tests
- Split ArmeriaKotlinRouteCollectorTest into four themed test classes
- Add fastTest and platformTest Gradle tasks for selective/parallel verification
- Wire suite tasks to reuse the IntelliJ Platform test classpath and JVM settings
