+++ plugin/src/test/kotlin/com/linecorp/intellij/plugins/armeria/explorer/ArmeriaKotlinSpringBootRouteCollectorTest.kt
@@ -1,13 +1,13 @@
 package com.linecorp.intellij.plugins.armeria.explorer
 
 import com.intellij.psi.search.GlobalSearchScope
-import com.intellij.testFramework.fixtures.LightJavaCodeInsightFixtureTestCase
+import com.linecorp.intellij.plugins.armeria.test.ArmeriaFixtureTestBase
 
-class ArmeriaKotlinSpringBootRouteCollectorTest : LightJavaCodeInsightFixtureTestCase() {
+class ArmeriaKotlinSpringBootRouteCollectorTest : ArmeriaFixtureTestBase() {
     override fun setUp() {
         super.setUp()
-        registerArmeriaStubs()
-        registerSpringBootStubs()
+        registerSpringAnnotationStubs()
+        registerArmeriaSpringStubs()
     }
 
     fun testCollectServiceRegistrationFromBeanConfigurator() {
@@ -439,69 +439,5 @@ class ArmeriaKotlinSpringBootRouteCollectorTest : LightJavaCodeInsightFixtureTes
         assertEquals(1, subtypeRoutes.size)
     }
 
-    private fun registerArmeriaStubs() {
-        myFixture.addClass(
-            """
-            package com.linecorp.armeria.server;
 
-            public final class Server {
-                public static ServerBuilder builder() {
-                    return null;
-                }
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package com.linecorp.armeria.server;
-
-            public final class ServerBuilder {
-                public ServerBuilder service(String path, Object service) {
-                    return this;
-                }
-
-                public ServerBuilder serviceUnder(String prefix, Object service) {
-                    return this;
-                }
-
-                public ServerBuilder annotatedService(Object service) {
-                    return this;
-                }
-
-                public Server build() {
-                    return null;
-                }
-            }
-            """.trimIndent(),
-        )
-    }
-
-    private fun registerSpringBootStubs() {
-        myFixture.addClass(
-            """
-            package org.springframework.context.annotation;
-
-            public @interface Bean {
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package org.springframework.context.annotation;
-
-            public @interface Configuration {
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package com.linecorp.armeria.spring;
-
-            @FunctionalInterface
-            public interface ArmeriaServerConfigurator {
-                void configure(com.linecorp.armeria.server.ServerBuilder serverBuilder);
-            }
-            """.trimIndent(),
-        )
-    }
 }
