+++ plugin/src/test/kotlin/com/linecorp/intellij/plugins/armeria/explorer/ArmeriaRouteDuplicateIndexTest.kt
@@ -1,12 +1,8 @@
 package com.linecorp.intellij.plugins.armeria.explorer
 
-import com.intellij.testFramework.fixtures.LightJavaCodeInsightFixtureTestCase
+import com.linecorp.intellij.plugins.armeria.test.ArmeriaFixtureTestBase
 
-class ArmeriaRouteDuplicateIndexTest : LightJavaCodeInsightFixtureTestCase() {
-    override fun setUp() {
-        super.setUp()
-        registerArmeriaStubs()
-    }
+class ArmeriaRouteDuplicateIndexTest : ArmeriaFixtureTestBase() {
 
     fun testDuplicateServiceRegistrationsInDifferentFilesAreReported() {
         myFixture.configureByText(
@@ -506,64 +502,4 @@ class ArmeriaRouteDuplicateIndexTest : LightJavaCodeInsightFixtureTestCase() {
         assertEquals(2, firstHits.single().registrationCount)
     }
 
-    private fun registerArmeriaStubs() {
-        myFixture.addClass(
-            """
-            package com.linecorp.armeria.server.annotation;
-
-            public @interface Get {
-                String value() default "";
-                String path() default "";
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package com.linecorp.armeria.server.annotation;
-
-            public @interface Post {
-                String value() default "";
-                String path() default "";
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package com.linecorp.armeria.server;
-
-            public final class Server {
-                public static ServerBuilder builder() {
-                    return null;
-                }
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package com.linecorp.armeria.server;
-
-            public final class ServerBuilder {
-                public ServerBuilder service(String path, Object service) {
-                    return this;
-                }
-
-                public ServerBuilder serviceUnder(String pathPrefix, Object service) {
-                    return this;
-                }
-
-                public ServerBuilder annotatedService(Object service) {
-                    return this;
-                }
-
-                public ServerBuilder annotatedService(String pathPrefix, Object service) {
-                    return this;
-                }
-
-                public com.linecorp.armeria.server.Server build() {
-                    return null;
-                }
-            }
-            """.trimIndent(),
-        )
-    }
 }
