+++ plugin/src/test/kotlin/com/linecorp/intellij/plugins/armeria/test/ArmeriaClientFixtureTestBase.kt
@@ -1,0 +1,111 @@
+package com.linecorp.intellij.plugins.armeria.test
+
+abstract class ArmeriaClientFixtureTestBase : ArmeriaFixtureTestBase() {
+
+    override fun registerArmeriaStubs() {
+        // Client tests do not need route collector stubs.
+    }
+
+    override fun setUp() {
+        super.setUp()
+        registerArmeriaClientStubs()
+    }
+
+    protected fun registerArmeriaClientStubs() {
+        myFixture.addClass(
+            """
+            package com.linecorp.armeria.client;
+
+            public final class WebClient {
+                public static WebClient of(String uri) {
+                    return null;
+                }
+
+                public static WebClientBuilder builder() {
+                    return null;
+                }
+
+                public static WebClientBuilder builder(String uri) {
+                    return null;
+                }
+            }
+            """.trimIndent(),
+        )
+        myFixture.addClass(
+            """
+            package com.linecorp.armeria.client;
+
+            public final class WebClientBuilder {
+            }
+            """.trimIndent(),
+        )
+        myFixture.addClass(
+            """
+            package com.linecorp.armeria.client.grpc;
+
+            public final class GrpcClient {
+                public static GrpcClientBuilder builder() {
+                    return null;
+                }
+            }
+            """.trimIndent(),
+        )
+        myFixture.addClass(
+            """
+            package com.linecorp.armeria.client.grpc;
+
+            public final class GrpcClientBuilder {
+            }
+            """.trimIndent(),
+        )
+        myFixture.addClass(
+            """
+            package com.linecorp.armeria.client.grpc;
+
+            public final class GrpcClients {
+                public static GrpcClientBuilder builder(String uri) {
+                    return null;
+                }
+
+                public static Object newClient(String uri, Class<?> stubClass) {
+                    return null;
+                }
+            }
+            """.trimIndent(),
+        )
+        myFixture.addClass(
+            """
+            package com.linecorp.armeria.client.thrift;
+
+            public final class ThriftClient {
+                public static ThriftClientBuilder builder() {
+                    return null;
+                }
+            }
+            """.trimIndent(),
+        )
+        myFixture.addClass(
+            """
+            package com.linecorp.armeria.client.thrift;
+
+            public final class ThriftClientBuilder {
+            }
+            """.trimIndent(),
+        )
+        myFixture.addClass(
+            """
+            package com.linecorp.armeria.client.thrift;
+
+            public final class ThriftClients {
+                public static ThriftClientBuilder builder(String uri) {
+                    return null;
+                }
+
+                public static Object newClient(String uri, Class<?> ifaceClass) {
+                    return null;
+                }
+            }
+            """.trimIndent(),
+        )
+    }
+}
